implement "link" operation
Our asymmetric ratchet scheme is not key private, and as such, we can
easily link public keys to their ciphertexts. This is done by using the
pairing: Even if we cannot compute the random value "s" in the
ciphertext, we can use the bilinearity of the pairing to see if the
public key matches.