implement more space-efficient ratchet
Previously, we took a random Gt element as basis for our key, and then
encrypted it with the HIBE. This allowed the recipient to decrypt the
rnadom element, derive the same key and finally apply the symmetric
decryption for the message.

That however is inefficient, as the Gt element is by far the biggest
element in our ciphertext (576 bytes). Additionally, there were
practical implementation problems, because it is the only element that
does not expose a `to_bytes`/`from_bytes` method or similar.

Now, what we do instead is to treat the HIBE more like a
"Diffie-Hellman" like key exchange: The sender picks s randomly and can
calculate e(g1, g2)^s, and the recipient can derive it from the other
two ciphertext components and his private key. Therefore, we don't pick
a message, and we simply don't sent the first ciphertext element, and
instead treat e(g1, g2)^s as the basis for our key derivation.