fix android serialization
There was an addition serialization step involved in the keys: We
serialized them in the method, and then again in
`serialize_to_bytearray`. That was of course nonsense, because when
deserializing the bytearray, we expect it to contain the serialization
of the key - and not the serialization of a vector containing the
serialization of the key! This added a few bytes in front of each key
(denoting the serialization length), which of course meant that the
actual deserialization would fail.

The fix is to pass the key itself to `serialize_to_bytearray`, easy
peasy!